Isolate SingleFileMetastore unit tests as an independent test suite in Metastore. (#394)
* Isolate SingleFileMetastore unit tests as an independent test suite in Metastore.
