+++ quickwit-metastore/src/metastore/single_file_metastore.rs
@@ -409,20 +409,14 @@ impl Metastore for SingleFileMetastore {
 
 #[cfg(test)]
 mod tests {
-    use std::collections::HashSet;
-    use std::ops::{Range, RangeInclusive};
     use std::path::Path;
     use std::sync::Arc;
-    use std::time::Duration;
-
-    use chrono::Utc;
-    use tokio::time::sleep;
 
     use quickwit_index_config::AllFlattenIndexConfig;
     use quickwit_storage::{MockStorage, StorageErrorKind};
 
-    use crate::{IndexMetadata, MetastoreError};
-    use crate::{Metastore, SingleFileMetastore, SplitMetadata, SplitState};
+    use crate::tests::*;
+    use crate::{IndexMetadata, Metastore, MetastoreError, SingleFileMetastore};
 
     #[tokio::test]
     async fn test_single_file_metastore_index_exists() {
@@ -448,42 +442,6 @@ mod tests {
             let result = metastore.index_exists(index_id).await.unwrap();
             let expected = true;
             assert_eq!(result, expected);
-        }
-    }
-
-    #[tokio::test]
-    async fn test_single_file_metastore_create_index() {
-        let metastore = SingleFileMetastore::for_test();
-        let index_id = "my-index";
-
-        {
-            // Check for the existence of index.
-            let result = metastore.index_exists(index_id).await.unwrap();
-            let expected = false;
-            assert_eq!(result, expected);
-
-            let index_metadata = IndexMetadata {
-                index_id: index_id.to_string(),
-                index_uri: "ram://indexes/my-index".to_string(),
-                index_config: Arc::new(AllFlattenIndexConfig::default()),
-            };
-
-            // Create index
-            metastore
-                .create_index(index_metadata.clone())
-                .await
-                .unwrap();
-
-            // Check for the existence of index.
-            let result = metastore.index_exists(index_id).await.unwrap();
-            let expected = true;
-            assert_eq!(result, expected);
-
-            let metastore_error = metastore.create_index(index_metadata).await.unwrap_err();
-            assert!(matches!(
-                metastore_error,
-                MetastoreError::IndexAlreadyExists { .. }
-            ));
         }
     }
 
@@ -538,1161 +496,63 @@ mod tests {
     }
 
     #[tokio::test]
+    async fn test_single_file_metastore_create_index() {
+        let metastore = SingleFileMetastore::for_test();
+        test_metastore_create_index(&metastore).await;
+    }
+
+    #[tokio::test]
     async fn test_single_file_metastore_delete_index() {
         let metastore = SingleFileMetastore::for_test();
-        let index_id = "my-index";
+        test_metastore_delete_index(&metastore).await;
+    }
 
-        {
-            // Check for the existence of index.
-            let result = metastore.index_exists(index_id).await.unwrap();
-            let expected = false;
-            assert_eq!(result, expected);
-
-            let index_metadata = IndexMetadata {
-                index_id: index_id.to_string(),
-                index_uri: "ram://indexes/my-index".to_string(),
-                index_config: Arc::new(AllFlattenIndexConfig::default()),
-            };
-
-            // Create index
-            metastore.create_index(index_metadata).await.unwrap();
-
-            // Check for the existence of index.
-            let result = metastore.index_exists(index_id).await.unwrap();
-            let expected = true;
-            assert_eq!(result, expected);
-
-            // Delete index
-            metastore.delete_index(index_id).await.unwrap();
-
-            // Delete a non-existent index.
-            let metastore_error = metastore
-                .delete_index("non-existent-index")
-                .await
-                .unwrap_err();
-            assert!(matches!(
-                metastore_error,
-                MetastoreError::IndexDoesNotExist { .. }
-            ));
-        }
+    #[tokio::test]
+    async fn test_single_file_metastore_index_metadata() {
+        let metastore = SingleFileMetastore::for_test();
+        test_metastore_index_metadata(&metastore).await;
     }
 
     #[tokio::test]
     async fn test_single_file_metastore_stage_split() {
         let metastore = SingleFileMetastore::for_test();
-        let index_id = "my-index";
-        let split_id = "one";
-        let split_metadata = SplitMetadata {
-            split_id: split_id.to_string(),
-            split_state: SplitState::Staged,
-            num_records: 1,
-            size_in_bytes: 2,
-            time_range: Some(RangeInclusive::new(0, 99)),
-            generation: 3,
-            update_timestamp: Utc::now().timestamp(),
-        };
-
-        {
-            // Check for the existence of index.
-            let result = metastore.index_exists(index_id).await.unwrap();
-            let expected = false;
-            assert_eq!(result, expected);
-
-            let index_metadata = IndexMetadata {
-                index_id: index_id.to_string(),
-                index_uri: "ram://indexes/my-index".to_string(),
-                index_config: Arc::new(AllFlattenIndexConfig::default()),
-            };
-
-            // Create index
-            metastore.create_index(index_metadata).await.unwrap();
-
-            // Check for the existence of index.
-            let result = metastore.index_exists(index_id).await.unwrap();
-            let expected = true;
-            assert_eq!(result, expected);
-
-            // stage split
-            metastore
-                .stage_split(index_id, split_metadata.clone())
-                .await
-                .unwrap();
-        }
-
-        {
-            let cache = metastore.cache.read().await;
-            assert_eq!(cache.get(index_id).unwrap().splits.len(), 1);
-            assert_eq!(
-                cache
-                    .get(index_id)
-                    .unwrap()
-                    .splits
-                    .get(split_id)
-                    .unwrap()
-                    .split_id,
-                "one".to_string()
-            );
-            assert_eq!(
-                cache
-                    .get(index_id)
-                    .unwrap()
-                    .splits
-                    .get(split_id)
-                    .unwrap()
-                    .split_state,
-                SplitState::Staged
-            );
-            assert_eq!(
-                cache
-                    .get(index_id)
-                    .unwrap()
-                    .splits
-                    .get(split_id)
-                    .unwrap()
-                    .num_records,
-                1
-            );
-            assert_eq!(
-                cache
-                    .get(index_id)
-                    .unwrap()
-                    .splits
-                    .get(split_id)
-                    .unwrap()
-                    .size_in_bytes,
-                2
-            );
-            assert_eq!(
-                cache
-                    .get(index_id)
-                    .unwrap()
-                    .splits
-                    .get(split_id)
-                    .unwrap()
-                    .time_range,
-                Some(RangeInclusive::new(0, 99))
-            );
-            assert_eq!(
-                cache
-                    .get(index_id)
-                    .unwrap()
-                    .splits
-                    .get(split_id)
-                    .unwrap()
-                    .generation,
-                3
-            );
-        }
-
-        {
-            // stage split (existing split id)
-            let metastore_error = metastore
-                .stage_split(index_id, split_metadata.clone())
-                .await
-                .unwrap_err();
-            assert!(matches!(
-                metastore_error,
-                MetastoreError::InternalError { .. }
-            ));
-        }
-
-        {
-            // stage split (non-existent index uri)
-            let metastore_error = metastore
-                .stage_split("non-existent-index", split_metadata.clone())
-                .await
-                .unwrap_err();
-            assert!(matches!(
-                metastore_error,
-                MetastoreError::IndexDoesNotExist { .. }
-            ));
-        }
+        test_metastore_stage_split(&metastore).await;
     }
 
     #[tokio::test]
-    async fn test_single_file_metastore_publish_split() {
+    async fn test_single_file_metastore_publish_splits() {
         let metastore = SingleFileMetastore::for_test();
-        let index_id = "my-index";
-        let split_id_one = "one";
-        let split_id_two = "two";
-        let current_timestamp = Utc::now().timestamp();
-        let split_metadata_one = SplitMetadata {
-            split_id: split_id_one.to_string(),
-            split_state: SplitState::Staged,
-            num_records: 1,
-            size_in_bytes: 2,
-            time_range: Some(RangeInclusive::new(0, 99)),
-            generation: 3,
-            update_timestamp: current_timestamp,
-        };
-        let split_metadata_two = SplitMetadata {
-            split_id: split_id_two.to_string(),
-            split_state: SplitState::Staged,
-            num_records: 5,
-            size_in_bytes: 6,
-            time_range: Some(RangeInclusive::new(30, 99)),
-            generation: 2,
-            update_timestamp: current_timestamp,
-        };
+        test_metastore_publish_splits(&metastore).await;
+    }
 
-        {
-            // Check for the existence of index.
-            let result = metastore.index_exists(index_id).await.unwrap();
-            let expected = false;
-            assert_eq!(result, expected);
+    #[tokio::test]
+    async fn test_single_file_metastore_mark_splits_as_deleted() {
+        let metastore = SingleFileMetastore::for_test();
+        test_metastore_mark_splits_as_deleted(&metastore).await;
+    }
 
-            let index_metadata = IndexMetadata {
-                index_id: index_id.to_string(),
-                index_uri: "ram://indexes/my-index".to_string(),
-                index_config: Arc::new(AllFlattenIndexConfig::default()),
-            };
+    #[tokio::test]
+    async fn test_single_file_metastore_delete_splits() {
+        let metastore = SingleFileMetastore::for_test();
+        test_metastore_delete_splits(&metastore).await;
+    }
 
-            // Create index
-            metastore.create_index(index_metadata).await.unwrap();
-
-            // Check for the existence of index.
-            let result = metastore.index_exists(index_id).await.unwrap();
-            let expected = true;
-            assert_eq!(result, expected);
-
-            // stage split
-            metastore
-                .stage_split(index_id, split_metadata_one.clone())
-                .await
-                .unwrap();
-
-            // publish split
-            metastore
-                .publish_splits(index_id, &[split_id_one])
-                .await
-                .unwrap();
-        }
-
-        {
-            let cache = metastore.cache.read().await;
-            assert_eq!(
-                cache
-                    .get(index_id)
-                    .unwrap()
-                    .splits
-                    .get(split_id_one)
-                    .unwrap()
-                    .split_state,
-                SplitState::Published
-            );
-        }
-
-        {
-            // publish published split
-            metastore
-                .publish_splits(index_id, &[split_id_one])
-                .await
-                .unwrap();
-
-            // publish non-staged split
-            metastore
-                .mark_splits_as_deleted(index_id, &[split_id_one]) // mark as deleted
-                .await
-                .unwrap();
-            let metastore_error = metastore
-                .publish_splits(index_id, &[split_id_one]) // publish
-                .await
-                .unwrap_err();
-            assert!(matches!(
-                metastore_error,
-                MetastoreError::SplitIsNotStaged { .. }
-            ));
-
-            // publish non-existent index
-            let metastore_error = metastore
-                .publish_splits("non-existent-index", &[split_id_one])
-                .await
-                .unwrap_err();
-            assert!(matches!(
-                metastore_error,
-                MetastoreError::IndexDoesNotExist { .. }
-            ));
-
-            // publish non-existent split
-            let metastore_error = metastore
-                .publish_splits(index_id, &["non-existent-split"])
-                .await
-                .unwrap_err();
-            assert!(matches!(
-                metastore_error,
-                MetastoreError::SplitDoesNotExist { .. }
-            ));
-        }
-
-        {
-            // publish one non-staged split and one non-existent split
-            let metastore_error = metastore
-                .publish_splits(index_id, &[split_id_one, split_id_two])
-                .await
-                .unwrap_err();
-            assert!(matches!(
-                metastore_error,
-                MetastoreError::SplitIsNotStaged { split_id: _ }
-            ));
-
-            // publish two non-existent splits
-            metastore
-                .delete_splits(index_id, &[split_id_one])
-                .await
-                .unwrap();
-            let metastore_error = metastore
-                .publish_splits(index_id, &[split_id_one, split_id_two])
-                .await
-                .unwrap_err();
-            assert!(matches!(
-                metastore_error,
-                MetastoreError::SplitDoesNotExist { split_id: _ }
-            ));
-
-            // publish one staged split and one non-exitent split
-            metastore
-                .stage_split(index_id, split_metadata_one.clone())
-                .await
-                .unwrap();
-            let metastore_error = metastore
-                .publish_splits(index_id, &[split_id_one, split_id_two])
-                .await
-                .unwrap_err();
-            assert!(matches!(
-                metastore_error,
-                MetastoreError::SplitDoesNotExist { .. }
-            ));
-
-            // publish two staged splits
-            metastore
-                .stage_split(index_id, split_metadata_two.clone())
-                .await
-                .unwrap();
-            metastore
-                .publish_splits(index_id, &[split_id_one, split_id_two])
-                .await
-                .unwrap();
-
-            //publishe two published splits
-            metastore
-                .publish_splits(index_id, &[split_id_one, split_id_two])
-                .await
-                .unwrap();
-        }
+    #[tokio::test]
+    async fn test_single_file_metastore_list_all_splits() {
+        let metastore = SingleFileMetastore::for_test();
+        test_metastore_list_all_splits(&metastore).await;
     }
 
     #[tokio::test]
     async fn test_single_file_metastore_list_splits() {
         let metastore = SingleFileMetastore::for_test();
-        let index_id = "my-index";
-        let current_timestamp = Utc::now().timestamp();
-
-        {
-            let index_metadata = IndexMetadata {
-                index_id: index_id.to_string(),
-                index_uri: "ram://indexes/my-index".to_string(),
-                index_config: Arc::new(AllFlattenIndexConfig::default()),
-            };
-
-            // create index
-            metastore.create_index(index_metadata).await.unwrap();
-        }
-
-        {
-            // stage split
-            let split_metadata_1 = SplitMetadata {
-                split_id: "one".to_string(),
-                split_state: SplitState::Staged,
-                num_records: 1,
-                size_in_bytes: 2,
-                time_range: Some(RangeInclusive::new(0, 99)),
-                generation: 3,
-                update_timestamp: current_timestamp,
-            };
-
-            let split_metadata_2 = SplitMetadata {
-                split_id: "two".to_string(),
-                split_state: SplitState::Staged,
-                num_records: 1,
-                size_in_bytes: 2,
-                time_range: Some(RangeInclusive::new(100, 199)),
-                generation: 3,
-                update_timestamp: current_timestamp,
-            };
-
-            let split_metadata_3 = SplitMetadata {
-                split_id: "three".to_string(),
-                split_state: SplitState::Staged,
-                num_records: 1,
-                size_in_bytes: 2,
-                time_range: Some(RangeInclusive::new(200, 299)),
-                generation: 3,
-                update_timestamp: current_timestamp,
-            };
-
-            let split_metadata_4 = SplitMetadata {
-                split_id: "four".to_string(),
-                split_state: SplitState::Staged,
-                num_records: 1,
-                size_in_bytes: 2,
-                time_range: Some(RangeInclusive::new(300, 399)),
-                generation: 3,
-                update_timestamp: current_timestamp,
-            };
-            let split_metadata_5 = SplitMetadata {
-                split_id: "five".to_string(),
-                split_state: SplitState::Staged,
-                num_records: 1,
-                size_in_bytes: 2,
-                time_range: None,
-                generation: 3,
-                update_timestamp: current_timestamp,
-            };
-
-            metastore
-                .stage_split(index_id, split_metadata_1)
-                .await
-                .unwrap();
-            metastore
-                .stage_split(index_id, split_metadata_2)
-                .await
-                .unwrap();
-            metastore
-                .stage_split(index_id, split_metadata_3)
-                .await
-                .unwrap();
-            metastore
-                .stage_split(index_id, split_metadata_4)
-                .await
-                .unwrap();
-            metastore
-                .stage_split(index_id, split_metadata_5)
-                .await
-                .unwrap();
-        }
-
-        {
-            // list
-            let time_range_opt = Some(Range {
-                start: 0i64,
-                end: 99i64,
-            });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, time_range_opt)
-                .await
-                .unwrap();
-            let split_ids: HashSet<String> = splits
-                .into_iter()
-                .map(|split_metadata| split_metadata.split_id)
-                .collect();
-            assert_eq!(split_ids.contains("one"), true);
-            assert_eq!(split_ids.contains("two"), false);
-            assert_eq!(split_ids.contains("three"), false);
-            assert_eq!(split_ids.contains("four"), false);
-            assert_eq!(split_ids.contains("five"), true);
-        }
-
-        {
-            // list
-            let time_range_opt = Some(Range {
-                start: 200,
-                end: i64::MAX,
-            });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, time_range_opt)
-                .await
-                .unwrap();
-            let split_ids: HashSet<String> = splits
-                .into_iter()
-                .map(|split_metadata| split_metadata.split_id)
-                .collect();
-            assert_eq!(split_ids.contains("one"), false);
-            assert_eq!(split_ids.contains("two"), false);
-            assert_eq!(split_ids.contains("three"), true);
-            assert_eq!(split_ids.contains("four"), true);
-            assert_eq!(split_ids.contains("five"), true);
-        }
-
-        {
-            // list
-            let time_range_opt = Some(Range {
-                start: i64::MIN,
-                end: 200,
-            });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, time_range_opt)
-                .await
-                .unwrap();
-            let split_ids: HashSet<String> = splits
-                .into_iter()
-                .map(|split_metadata| split_metadata.split_id)
-                .collect();
-            assert_eq!(split_ids.contains("one"), true);
-            assert_eq!(split_ids.contains("two"), true);
-            assert_eq!(split_ids.contains("three"), false);
-            assert_eq!(split_ids.contains("four"), false);
-            assert_eq!(split_ids.contains("five"), true);
-        }
-
-        {
-            // list
-            let range = Some(Range { start: 0, end: 100 });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = Some(Range { start: 0, end: 101 });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = Some(Range { start: 0, end: 199 });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = Some(Range { start: 0, end: 200 });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = Some(Range { start: 0, end: 201 });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = Some(Range { start: 0, end: 299 });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = Some(Range { start: 0, end: 300 });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = Some(Range { start: 0, end: 301 });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = Some(Range {
-                start: 301,
-                end: 400,
-            });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = Some(Range {
-                start: 300,
-                end: 400,
-            });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = Some(Range {
-                start: 299,
-                end: 400,
-            });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = Some(Range {
-                start: 201,
-                end: 400,
-            });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = Some(Range {
-                start: 200,
-                end: 400,
-            });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = Some(Range {
-                start: 199,
-                end: 400,
-            });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = Some(Range {
-                start: 101,
-                end: 400,
-            });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = Some(Range {
-                start: 101,
-                end: 400,
-            });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = Some(Range {
-                start: 100,
-                end: 400,
-            });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = Some(Range {
-                start: 99,
-                end: 400,
-            });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = Some(Range {
-                start: 1000,
-                end: 1100,
-            });
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), false);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-
-        {
-            // list
-            let range = None;
-            let splits = metastore
-                .list_splits(index_id, SplitState::Staged, range)
-                .await
-                .unwrap();
-            let mut split_id_vec = Vec::new();
-            for split_metadata in splits {
-                split_id_vec.push(split_metadata.split_id);
-            }
-            assert_eq!(split_id_vec.contains(&"one".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"two".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"three".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"four".to_string()), true);
-            assert_eq!(split_id_vec.contains(&"five".to_string()), true);
-        }
-    }
-
-    #[tokio::test]
-    async fn test_single_file_metastore_mark_split_as_deleted() {
-        let metastore = SingleFileMetastore::for_test();
-        let index_id = "my-index";
-        let split_id = "split-one";
-        let split_metadata = SplitMetadata {
-            split_id: split_id.to_string(),
-            split_state: SplitState::Staged,
-            num_records: 1,
-            size_in_bytes: 2,
-            time_range: Some(RangeInclusive::new(0, 99)),
-            generation: 3,
-            update_timestamp: Utc::now().timestamp(),
-        };
-
-        {
-            // Check for the existence of index.
-            let result = metastore.index_exists(index_id).await.unwrap();
-            let expected = false;
-            assert_eq!(result, expected);
-
-            let index_metadata = IndexMetadata {
-                index_id: index_id.to_string(),
-                index_uri: "ram://indexes/my-index".to_string(),
-                index_config: Arc::new(AllFlattenIndexConfig::default()),
-            };
-
-            // Create index
-            metastore.create_index(index_metadata).await.unwrap();
-
-            // Check for the existence of index.
-            let result = metastore.index_exists(index_id).await.unwrap();
-            let expected = true;
-            assert_eq!(result, expected);
-
-            // stage split
-            metastore
-                .stage_split(index_id, split_metadata.clone())
-                .await
-                .unwrap();
-
-            // publish split
-            metastore
-                .publish_splits(index_id, &[split_id])
-                .await
-                .unwrap();
-
-            // mark splits as deleted (one non-existent)
-            let metastore_error = metastore
-                .mark_splits_as_deleted(index_id, &[split_id, "non-existent-split"])
-                .await
-                .unwrap_err();
-            assert!(matches!(
-                metastore_error,
-                MetastoreError::SplitDoesNotExist { .. }
-            ));
-
-            // mark split as deleted
-            metastore
-                .mark_splits_as_deleted(index_id, &[split_id])
-                .await
-                .unwrap();
-        }
-
-        {
-            let cache = metastore.cache.read().await;
-            assert_eq!(
-                cache
-                    .get(index_id)
-                    .unwrap()
-                    .splits
-                    .get(split_id)
-                    .unwrap()
-                    .split_state,
-                SplitState::ScheduledForDeletion
-            );
-        }
-
-        {
-            // mark split as deleted (already marked as deleted)
-            metastore
-                .mark_splits_as_deleted(index_id, &[split_id])
-                .await
-                .unwrap();
-
-            // mark split as deleted (non-existent index)
-            let metastore_error = metastore
-                .mark_splits_as_deleted("non-existent-index", &[split_id])
-                .await
-                .unwrap_err();
-            assert!(matches!(
-                metastore_error,
-                MetastoreError::IndexDoesNotExist { .. }
-            ));
-
-            // mark split as deleted (non-existent)
-            let metastore_error = metastore
-                .mark_splits_as_deleted(index_id, &["non-existent-split"])
-                .await
-                .unwrap_err();
-            assert!(matches!(
-                metastore_error,
-                MetastoreError::SplitDoesNotExist { .. }
-            ));
-        }
-    }
-
-    #[tokio::test]
-    async fn test_single_file_metastore_delete_split() {
-        let metastore = SingleFileMetastore::for_test();
-        let index_id = "my-index";
-        let split_id = "split-one";
-        let split_metadata = SplitMetadata {
-            split_id: split_id.to_string(),
-            split_state: SplitState::Staged,
-            num_records: 1,
-            size_in_bytes: 2,
-            time_range: Some(RangeInclusive::new(0, 100)),
-            generation: 3,
-            update_timestamp: Utc::now().timestamp(),
-        };
-
-        {
-            // Check for the existence of index.
-            let result = metastore.index_exists(index_id).await.unwrap();
-            assert_eq!(result, false);
-
-            let index_metadata = IndexMetadata {
-                index_id: index_id.to_string(),
-                index_uri: "ram://indexes/my-index".to_string(),
-                index_config: Arc::new(AllFlattenIndexConfig::default()),
-            };
-
-            // Create index
-            metastore.create_index(index_metadata).await.unwrap();
-
-            // Check for the existence of index.
-            let result = metastore.index_exists(index_id).await.unwrap();
-            assert_eq!(result, true);
-
-            // stage split
-            metastore
-                .stage_split(index_id, split_metadata.clone())
-                .await
-                .unwrap();
-
-            // publish split
-            metastore
-                .publish_splits(index_id, &[split_id])
-                .await
-                .unwrap();
-
-            // delete split (published split)
-            let metastore_error = metastore
-                .delete_splits(index_id, &[split_id])
-                .await
-                .unwrap_err();
-            assert!(matches!(metastore_error, MetastoreError::Forbidden { .. }));
-
-            // mark split as deleted
-            metastore
-                .mark_splits_as_deleted(index_id, &[split_id])
-                .await
-                .unwrap();
-
-            // mark splits as deleted (one non-existent split)
-            let metastore_error = metastore
-                .mark_splits_as_deleted(index_id, &[split_id, "non-existent-split"])
-                .await
-                .unwrap_err();
-            assert!(matches!(
-                metastore_error,
-                MetastoreError::SplitDoesNotExist { .. }
-            ));
-
-            // delete split
-            metastore
-                .delete_splits(index_id, &[split_id])
-                .await
-                .unwrap();
-        }
-
-        {
-            let cache = metastore.cache.read().await;
-            assert_eq!(
-                cache.get(index_id).unwrap().splits.contains_key(split_id),
-                false
-            );
-        }
-
-        {
-            // mark split as deleted (non-existent index)
-            let metastore_error = metastore
-                .delete_splits("non-existent-index", &[split_id])
-                .await
-                .unwrap_err();
-            assert!(matches!(
-                metastore_error,
-                MetastoreError::IndexDoesNotExist { .. }
-            ));
-
-            // delete split (non-existent split)
-            let metastore_error = metastore
-                .delete_splits(index_id, &["non-existent-split"])
-                .await
-                .unwrap_err();
-            assert!(matches!(
-                metastore_error,
-                MetastoreError::SplitDoesNotExist { .. }
-            ));
-        }
+        test_metastore_list_splits(&metastore).await;
     }
 
     #[tokio::test]
     async fn test_single_file_metastore_split_update_timestamp() {
         let metastore = SingleFileMetastore::for_test();
-        let index_id = "my-index";
-        let split_id = "split-one";
-        let mut current_timestamp = Utc::now().timestamp();
-        let split_metadata = SplitMetadata {
-            split_id: split_id.to_string(),
-            split_state: SplitState::New,
-            num_records: 1,
-            size_in_bytes: 2,
-            time_range: Some(RangeInclusive::new(0, 99)),
-            generation: 3,
-            update_timestamp: current_timestamp,
-        };
-        let index_metadata = IndexMetadata {
-            index_id: index_id.to_string(),
-            index_uri: "ram://indexes/my-index".to_string(),
-            index_config: Arc::new(AllFlattenIndexConfig::default()),
-        };
-
-        // Create index
-        metastore.create_index(index_metadata).await.unwrap();
-
-        // wait for 1s, stage split & check `update_timestamp`
-        sleep(Duration::from_secs(1)).await;
-        metastore
-            .stage_split(index_id, split_metadata.clone())
-            .await
-            .unwrap();
-        let split_meta = metastore.list_all_splits(index_id).await.unwrap()[0].clone();
-        assert!(split_meta.update_timestamp > current_timestamp);
-        current_timestamp = split_meta.update_timestamp;
-
-        // wait for 1s, publish split & check `update_timestamp`
-        sleep(Duration::from_secs(1)).await;
-        metastore
-            .publish_splits(index_id, &[split_id])
-            .await
-            .unwrap();
-        let split_meta = metastore.list_all_splits(index_id).await.unwrap()[0].clone();
-        assert!(split_meta.update_timestamp > current_timestamp);
-        current_timestamp = split_meta.update_timestamp;
-
-        // wait for 1s, mark split as deleted & check `update_timestamp`
-        sleep(Duration::from_secs(1)).await;
-        metastore
-            .mark_splits_as_deleted(index_id, &[split_id])
-            .await
-            .unwrap();
-        let split_meta = metastore.list_all_splits(index_id).await.unwrap()[0].clone();
-        assert!(split_meta.update_timestamp > current_timestamp);
+        test_metastore_split_update_timestamp(&metastore).await;
     }
 
     #[tokio::test]
@@ -1713,50 +573,6 @@ mod tests {
         });
 
         let metastore = SingleFileMetastore::new(Arc::new(mock_storage));
-
-        let index_id = "my-index";
-        let split_id = "split-one";
-        let split_metadata = SplitMetadata {
-            split_id: split_id.to_string(),
-            split_state: SplitState::Staged,
-            num_records: 1,
-            size_in_bytes: 2,
-            time_range: None,
-            generation: 3,
-            update_timestamp: Utc::now().timestamp(),
-        };
-
-        let index_metadata = IndexMetadata {
-            index_id: index_id.to_string(),
-            index_uri: "ram://my-indexes/my-index".to_string(),
-            index_config: Arc::new(AllFlattenIndexConfig::default()),
-        };
-
-        // create index
-        metastore.create_index(index_metadata).await.unwrap();
-
-        // stage split
-        metastore
-            .stage_split(index_id, split_metadata)
-            .await
-            .unwrap();
-
-        // publish split fails
-        let err = metastore.publish_splits(index_id, &[split_id]).await;
-        assert!(err.is_err());
-
-        // empty
-        let split = metastore
-            .list_splits(index_id, SplitState::Published, None)
-            .await
-            .unwrap();
-        assert!(split.is_empty());
-
-        // not empty
-        let split = metastore
-            .list_splits(index_id, SplitState::Staged, None)
-            .await
-            .unwrap();
-        assert!(!split.is_empty());
+        test_metastore_storage_failing(&metastore).await;
     }
 }
